fixed ubuntu 16

+++ install.sh
@@ -59,12 +59,12 @@ then
         else
             read -p "install python${PYSETENV_PYTHON_VERSION} on the system (Y/N)" y_n
             case $y_n in
-                Y|y 
+                Y|y) 
                     add-apt-repository ppa:fkrull/deadsnakes
                     apt-get update
                     apt-get install python${PYSETENV_PYTHON_VERSION}
                     apt-get autoremove -y ;;
-                N|n echo "[!] Aborting";;
+                N|n) echo "[!] Aborting";;
                 *) echo "[*] Enter either Y|y for yes or N|n for no"
             esac
         fi
